+++ src/aux.rs
@@ -307,6 +307,22 @@ impl<T: AsRef<StblBox>> SampleAccessor<'_, T> {
         stss_box.sample_numbers.binary_search(&self.index).is_ok()
     }
 
+    /// このサンプルをデコードするために必要となる同期サンプルのインデックスを返す
+    ///
+    /// 自分自身が同期サンプルの場合には、自分のインデックスが返される。
+    /// 自分よりも前方に同期サンプルが存在しない場合には [`None`] が返される。
+    pub fn sync_sample_index(&self) -> Option<NonZeroU32> {
+        let Some(stss_box) = &self.sample_table.stbl_box().stss_box else {
+            return Some(self.index);
+        };
+
+        match stss_box.sample_numbers.binary_search(&self.index) {
+            Ok(_) => Some(self.index),
+            Err(0) => None,
+            Err(i) => Some(stss_box.sample_numbers[i - 1]),
+        }
+    }
+
     /// サンプルが属するチャンクの情報を返す
     pub fn chunk(&self) -> ChunkAccessor<T> {
         let i = self
@@ -437,6 +453,10 @@ mod tests {
             assert_eq!(sample.data_size(), i as u32 + 1);
             assert_eq!(sample.data_offset(), sample_offsets[i] as u64);
             assert_eq!(sample.is_sync_sample(), (i + 1) % 2 == 1);
+            assert_eq!(
+                sample.sync_sample_index(),
+                Some(NonZeroU32::MIN.saturating_add(i as u32 / 2 * 2))
+            );
             assert_eq!(sample.chunk().index().get(), sample_chunks[i]);
         }
         assert!(sample_table.get_sample(index(11)).is_none());